+++ CHANGELOG.md
@@ -24,6 +24,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/jeslage/adaptive-modular-scale/compare/v1.0.0...develop
+[unreleased]: https://github.com/jeslage/adaptive-modular-scale/compare/v1.0.1...develop
 [1.0.1]: https://github.com/jeslage/adaptive-modular-scale/compare/v1.0.0-rc.1...v1.0.1
 [1.0.0-rc.1]: https://github.com/jeslage/adaptive-modular-scale/compare/v1.0.0...v1.0.0-rc.1